Organized screen to make it a component, and showed component on app.js
